+++ examples/src/main/scala/mongo4s/examples/bsonDirectCodecs.scala
@@ -2,8 +2,6 @@ package mongo4s.examples
 
 import mongo4s.bson.direct.{ScalarWireCodec, WireCodec}
 
-import mongo4s.bson.BsonInstances.given
-
 object bsonDirectCodecs:
 
   given ScalarWireCodec[Role] = ScalarWireCodec[String].imap(Role.valueOf)(_.toString)